+++ src/SharpDeck/Connectivity/StreamDeckActionCacheCollection.cs
@@ -96,19 +96,9 @@ namespace SharpDeck.Connectivity
         /// </summary>
         /// <param name="key">The <see cref="IActionEventArgs" /> instance containing the event data.</param>
         /// <param name="action">The result action.</param>
-        /// <returns><c>true</c> when the action was found; otherwise <c>false</c>.</returns>
-        public bool TryGet(IActionEventArgs key, out StreamDeckAction action)
-            => this.TryGet<SettingsPayload>(key, out action, null);
-
-        /// <summary>
-        /// Attempts to get the cached <see cref="StreamDeckAction" /> for the specified <paramref name="key" />
-        /// </summary>
-        /// <param name="key">The <see cref="IActionEventArgs" /> instance containing the event data.</param>
-        /// <param name="action">The result action.</param>
         /// <param name="payload">The optional payload containing the settings; when these are supplied, the cache item is validated.</param>
         /// <returns><c>true</c> when the action was found, and valid if <paramref name="payload"/> were supplied; otherwise <c>false</c>.</returns>
-        public bool TryGet<T>(IActionEventArgs key, out StreamDeckAction action, T payload)
-            where T : SettingsPayload
+        public bool TryGet(IActionEventArgs key, out StreamDeckAction action, SettingsPayload payload = null)
         {
             try
             {